✨ feat: 콘테스트 마감 후 투표 결과, 수상자 선정, contest 상태 update

+++ src/main/java/com/mulmeong/batchserver/contest/entity/contestRead/ContestPostRead.java
@@ -25,7 +25,7 @@ public class ContestPostRead {
     private String mediaUrl;
     private MediaType mediaType;
     private LocalDateTime createdAt;
-    private Integer voteCount;
+    private Long voteCount;
 
 
 }
